windsurf-project/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
watchdog==2.1.6
pytesseract==0.3.10
numpy==2.3.4
pgsocr==0.1.5
pillow==12.0.0
tesserocr==2.9.1
tqdm==4.67.1
deep-translator==1.11.4
pysubs2==1.7.3
beautifulsoup4==4.12.3
retry==0.9.2
subliminal==2.2.1
guessit==3.8.0
babelfish==0.6.1
enzyme==0.5.2
requests==2.32.3
deepl==1.18.0
azure-ai-translation-text==1.0.0